updated test-code
